+++ src/app/app.config.ts
@@ -1,7 +1,7 @@
 import { ApplicationConfig } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
-import { CO2_CALCULATOR, CO2_CONFIG } from './facades/CO2InjectionToken';
-import { CO2Calculator } from './facades/CO2Calculator';
+import { CO2_CALCULATOR, CO2_CONFIG } from './services/facades/CO2InjectionToken';
+import { CO2Calculator } from './services/facades/CO2Calculator';
 
 export const appConfig: ApplicationConfig = {
   providers: [
