+++ src/app/test-utils/fake-co2-calculator.testing.ts
@@ -1,5 +1,5 @@
-import { CO2_CALCULATOR } from '../facades/CO2InjectionToken';
-import { FakeCO2Calculator } from '../facades/FakeCO2Calculator';
+import { CO2_CALCULATOR } from '../services/facades/CO2InjectionToken';
+import { FakeCO2Calculator } from '../services/facades/FakeCO2Calculator';
 
 export const fakeCO2CalculatorFactory = () => {
   const instance = new FakeCO2Calculator();
